pydantic-ai/pydantic-ai-agent-etl/requirements.txt
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
fastapi
faststream
aiokafka
uvicorn
python-json-logger
celery
redis
sse-starlette
prometheus-fastapi-instrumentator
saq[web,redis]
httpx
pydantic-ai==0.0.14
python-dotenv==1.0.1
rich==13.9.4
asyncio-mqtt==0.16.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp_basicauth==1.1.0
aiokafka==0.12.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
croniter==6.0.0
debugpy==1.8.11
fast-depends==2.4.12
fastapi==0.115.6
faststream==0.5.34
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
redis==5.2.1
saq==0.19.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
